reset mfa for the user

+++ src/components/ResourceTable/trustedDevices/actions/Delete/index.js
@@ -1,7 +1,8 @@
 import { useDispatch } from 'react-redux';
-import { useEffect } from 'react';
+import { useCallback } from 'react';
 import actions from '../../../../../actions';
 import TrashIcon from '../../../../icons/TrashIcon';
+import useConfirmDialog from '../../../../ConfirmDialog';
 
 export default {
   key: 'deleteDevice',
@@ -10,11 +11,28 @@ export default {
   useOnClick: rowData => {
     const { _id: deviceId} = rowData;
     const dispatch = useDispatch();
+    const { confirmDialog } = useConfirmDialog();
 
-    useEffect(() => {
-      dispatch(actions.mfa.deleteDevice(deviceId));
-    }, [dispatch, deviceId]);
+    const handleClick = useCallback(() => {
+      confirmDialog({
+        title: 'Delete trusted MFA device?',
+        message: "Are you sure you want to delete your trusted MFA device? You'll need to re-authenticate your account the next time you sign into integrator.io with the device.Are you sure you want to delete your trusted MFA device? You'll need to re-authenticate your account the next time you sign into integrator.io with the device.",
+        buttons: [
+          {
+            label: 'Delete',
+            error: true,
+            onClick: () => {
+              dispatch(actions.mfa.deleteDevice(deviceId));
+            },
+          },
+          {
+            label: 'Cancel',
+            variant: 'text',
+          },
+        ],
+      });
+    }, [confirmDialog, dispatch, deviceId]);
 
-    return null;
+    return handleClick;
   },
 };
